Add Calix B6 save_config and telnet support; correct exit_config_mode; add to device templates
