+++ netmiko/calix/__init__.py
@@ -1,4 +1,4 @@
 from __future__ import unicode_literals
-from netmiko.calix.calix_b6_ssh import CalixB6SSH
+from netmiko.calix.calix_b6 import CalixB6SSH, CalixB6Telnet
 
-__all__ = ['CalixB6SSH']
+__all__ = ['CalixB6SSH', 'CalixB6Telnet']
